parsers: revamp the interface of yytnamerr
Currently yytnamerr takes a string, and strips its quotes when
possible.  It is actually more convenient to take a token number and
to transform the token name.

* data/skeletons/glr.c, data/skeletons/lalr1.cc, data/skeletons/lalr1.d,
* data/skeletons/lalr1.java, data/skeletons/yacc.c
(yytnamerr): Take a token number instead of a string.
(yysyntax_error): Adjust: keep format arguments as a list of token
numbers, instead of a list of strings.